Moved from bika.lims also related to LIMS-11
