+++ src/bika/extras/browser/overrides/overrides.zcml
@@ -16,13 +16,6 @@
       layer="bika.lims.interfaces.IBikaLIMS"
     />
 
-    <browser:page
-      for="bika.lims.interfaces.IAnalysisProfiles"
-      name="folder_view"
-      class=".bika_analysisprofiles.AnalysisProfilesView"
-      permission="senaite.core.permissions.ManageBika"
-      layer="bika.lims.interfaces.IBikaLIMS"
-    />
 
     <browser:page
       for="bika.lims.interfaces.IAnalysisSpec"
